+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.8
 - Fixed item pile token image not being updated when first creating a new item pile
+- Fixed `Take All Items` button not working on items
 
 ## Version 1.0.7
 - Fixed module being broken